rust-openssh/src/addr.rs
use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
use std::ptr;

/// Address family constants, similar to C's `AF_INET`, `AF_INET6`
pub const AF_INET: u16 = 2;
pub const AF_INET6: u16 = 10;

#[repr(C)]
pub union XAddrUnion {
    pub v4: Ipv4Addr,
    pub v6: Ipv6Addr,
    addr32: [u32; 4],
}

#[repr(C)]
pub struct XAddr {
    pub af: u16,            // Address family (AF_INET or AF_INET6)
    pub xa: XAddrUnion,     // Union to store address (v4, v6 or addr32)
    pub scope_id: u32,      // For IPv6 scope ID (interface index)
}

impl XAddr {
    // Convert sockaddr to XAddr (this function would be used in the context of network-related operations)
    pub fn addr_sa_to_xaddr(sa: &std::net::SocketAddr) -> Option<XAddr> {
        let mut xaddr = XAddr {
            af: 0,
            xa: XAddrUnion { v4: Ipv4Addr::new(0, 0, 0, 0) },
            scope_id: 0,
        };
        
        match sa {
            std::net::SocketAddr::V4(addr) => {
                xaddr.af = AF_INET;
                xaddr.xa = XAddrUnion { v4: *addr.ip() };
            }
            std::net::SocketAddr::V6(addr) => {
                xaddr.af = AF_INET6;
                xaddr.xa = XAddrUnion { v6: *addr.ip() };
                xaddr.scope_id = addr.scope_id();
            }
        }
        Some(xaddr)
    }

    // Convert XAddr to a string representation of the address
    pub fn addr_ntop(&self) -> String {
        unsafe {
            match self.af {
                AF_INET => format!("{}", self.xa.v4),  // Accessing v4 inside unsafe block
                AF_INET6 => format!("{}", self.xa.v6), // Accessing v6 inside unsafe block
                _ => String::new(),
            }
        }
    }

    // Network mask length calculation for IPv4 and IPv6
    fn addr_unicast_masklen(af: u16) -> i32 {
        match af {
            AF_INET => 32,
            AF_INET6 => 128,
            _ => -1,
        }
    }

    // Validity check for masklen
    fn masklen_valid(af: u16, masklen: u32) -> i32 {
        match af {
            AF_INET => if masklen <= 32 { 0 } else { -1 },
            AF_INET6 => if masklen <= 128 { 0 } else { -1 },
            _ => -1,
        }
    }

    // Calculate the network mask for the given address family and length
    pub fn addr_netmask(af: u16, l: u32, n: &mut XAddr) -> i32 {
        if XAddr::masklen_valid(af, l) != 0 {
            return -1;
        }

        unsafe {
            // Initialize the address to zero
            std::ptr::write_bytes(n, 0, std::mem::size_of::<XAddr>());

            match af {
                AF_INET => {
                    n.af = AF_INET;
                    if l == 0 {
                        return 0;
                    }
                    // Creating netmask for IPv4
                    n.xa.addr32[0] = u32::MAX << (32 - l);
                }
                AF_INET6 => {
                    n.af = AF_INET6;
                    let mut i = 0;
                    let mut len = l;
                    while i < 4 && len >= 32 {
                        n.xa.addr32[i] = u32::MAX;
                        i += 1;
                        len -= 32;
                    }
                    if i < 4 && len != 0 {
                        n.xa.addr32[i] = u32::MAX << (32 - len);
                    }
                }
                _ => return -1,
            }
        }
        0
    }

    // Perform logical AND of two addresses and store the result in `dst`
    pub fn addr_and(dst: &mut XAddr, a: &XAddr, b: &XAddr) -> i32 {
        if a.af != b.af {
            return -1;
        }

        unsafe {
            std::ptr::copy_nonoverlapping(a, dst, std::mem::size_of::<XAddr>());
            match a.af {
                AF_INET => {
                    dst.xa.addr32[0] &= b.xa.addr32[0];
                }
                AF_INET6 => {
                    for i in 0..4 {
                        dst.xa.addr32[i] &= b.xa.addr32[i];
                    }
                }
                _ => return -1,
            }
        }

        0
    }

    // Increment the address (IPv4 or IPv6)
    pub fn addr_increment(&mut self) {
        unsafe {
            match self.af {
                AF_INET => {
                    // Increment the IPv4 address
                    let mut val = u32::from(self.xa.v4);
                    val += 1;
                    self.xa.v4 = Ipv4Addr::from(val);
                }
                AF_INET6 => {
                    // Increment the IPv6 address
                    for i in 0..4 {
                        let mut val = self.xa.addr32[i];
                        val += 1;
                        self.xa.addr32[i] = val;
                        if val != 0 {
                            break;
                        }
                    }
                }
                _ => {}
            }
        }
    }

    // Convert the string address to `XAddr`
    pub fn addr_pton(addr: &str) -> Option<XAddr> {
        let socket_addr: Option<std::net::SocketAddr> = addr.parse().ok();
        socket_addr.and_then(|sa| XAddr::addr_sa_to_xaddr(&sa))
    }
}

fn main() {
    // Example usage
    let addr = XAddr::addr_pton("192.168.1.1").unwrap();
    println!("Address: {}", addr.addr_ntop());

    let mut xaddr = XAddr {
        af: AF_INET,
        xa: XAddrUnion { v4: Ipv4Addr::new(0, 0, 0, 0) },
        scope_id: 0,
    };

    // Set a netmask for the address
    XAddr::addr_netmask(AF_INET, 24, &mut xaddr);
    println!("Netmask: {}", xaddr.addr_ntop());

    // Increment address
    xaddr.addr_increment();
    println!("Incremented address: {}", xaddr.addr_ntop());
}
